reconocimientofacial/requirements.txt
Flask==3.1.3
opencv-contrib-python==4.10.0.84
numpy>=1.20.0,<3.0.0
mysql-connector-python==9.6.0
flask-cors==6.0.2
python-dotenv